dom/base/test/jsmodules/test_moduleScriptsNotRun.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test script elements with type="module" are not run for content HTML</title>
<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var inlineModuleRan = false;
  var moduleRan = false;

  SimpleTest.waitForExplicitFinish();

  function testLoaded() {
    ok(!moduleRan, "Check module script was not run");
    ok(!inlineModuleRan, "Check inline module script was not run");
    SimpleTest.finish();
  }
</script>
<script type="module" src="module_setRan.js"></script>
<script type="module">inlineModuleRan = true;</script>
<body onload='testLoaded()'></body>
